Changed the architecture, built the login form from the registration form.
I added chunks for pages, tried to split the registration form into a standard template, and eventually created form_template.I also thought that footer. sccs header.sccs and other site components should be stored in the components folder, so that it is convenient to build sites directly as a constructor.